Filter summary in the bulk inspector
