+++ apps/webapp/app/routes/resources.orgs.$organizationSlug.projects.$projectParam.env.$envParam.runs.bulkaction.tsx
@@ -14,6 +14,12 @@ import { ExitIcon } from "~/assets/icons/ExitIcon";
 import selectRunsIndividually from "~/assets/images/select-runs-individually.png";
 import selectRunsUsingFilters from "~/assets/images/select-runs-using-filters.png";
 import {
+  BulkActionAction,
+  BulkActionFilterSummary,
+  BulkActionMode,
+  EstimatedCount,
+} from "~/components/BulkActionFilterSummary";
+import {
   Accordion,
   AccordionContent,
   AccordionItem,
@@ -92,11 +98,6 @@ export async function loader({ request, params }: LoaderFunctionArgs) {
 
   return typedjson(data);
 }
-
-const BulkActionMode = z.union([z.literal("selected"), z.literal("filter")]);
-type BulkActionMode = z.infer<typeof BulkActionMode>;
-const BulkActionAction = z.union([z.literal("cancel"), z.literal("replay")]);
-type BulkActionAction = z.infer<typeof BulkActionAction>;
 
 export const CreateBulkActionSearchParams = z.object({
   mode: BulkActionMode.default("filter"),
@@ -364,7 +365,7 @@ export function CreateBulkActionInspector({
             </InputGroup>
             <InputGroup>
               <Label>Preview</Label>
-              <BulkActionPreview
+              <BulkActionFilterSummary
                 selected={mode === "selected" ? selectedItems.size : data?.count}
                 mode={mode as BulkActionMode}
                 action={action as BulkActionAction}
@@ -401,7 +402,7 @@ export function CreateBulkActionInspector({
             <DialogContent>
               <DialogHeader>{action === "replay" ? "Replay runs" : "Cancel runs"}</DialogHeader>
               <div className="flex flex-col gap-3 divide-y divide-grid-dimmed pt-2">
-                <BulkActionPreview
+                <BulkActionFilterSummary
                   selected={mode === "selected" ? selectedItems.size : data?.count}
                   mode={mode as BulkActionMode}
                   action={action as BulkActionAction}
@@ -455,221 +456,4 @@ export function CreateBulkActionInspector({
       </div>
     </Form>
   );
-}
-
-function BulkActionPreview({
-  selected,
-  mode,
-  action,
-  filters,
-}: {
-  selected?: number;
-  mode: BulkActionMode;
-  action: BulkActionAction;
-  filters: TaskRunListSearchFilters;
-}) {
-  switch (mode) {
-    case "selected":
-      return (
-        <Paragraph variant="small">
-          You have individually selected {simplur`${selected} run[|s]`} to be{" "}
-          <Action action={action} />.
-        </Paragraph>
-      );
-    case "filter": {
-      const { label, valueLabel, rangeType } = timeFilterRenderValues({
-        from: filters.from ? dateFromString(`${filters.from}`) : undefined,
-        to: filters.to ? dateFromString(`${filters.to}`) : undefined,
-        period: filters.period,
-      });
-
-      return (
-        <div className="flex flex-col gap-2">
-          <Paragraph variant="small">
-            You have selected{" "}
-            <span className="text-text-bright">
-              <EstimatedCount count={selected} />
-            </span>{" "}
-            runs to be <Action action={action} /> using these filters:
-          </Paragraph>
-          <div className="flex flex-col gap-2">
-            <AppliedFilter
-              variant="minimal/medium"
-              label={label}
-              icon={filterIcon("period")}
-              value={valueLabel}
-              removable={false}
-            />
-            {Object.entries(filters).map(([key, value]) => {
-              if (!value && key !== "period") {
-                return null;
-              }
-
-              const typedKey = key as TaskRunListSearchFilterKey;
-
-              switch (typedKey) {
-                case "cursor":
-                case "direction":
-                case "environments":
-                //We need to handle time differently because we have a default
-                case "period":
-                case "from":
-                case "to": {
-                  return null;
-                }
-                case "tasks": {
-                  const values = Array.isArray(value) ? value : [`${value}`];
-                  return (
-                    <AppliedFilter
-                      variant="minimal/medium"
-                      key={key}
-                      label={filterTitle(key)}
-                      icon={filterIcon(key)}
-                      value={appliedSummary(values)}
-                      removable={false}
-                    />
-                  );
-                }
-                case "versions": {
-                  const values = Array.isArray(value) ? value : [`${value}`];
-                  return (
-                    <AppliedFilter
-                      variant="minimal/medium"
-                      key={key}
-                      label={filterTitle(key)}
-                      icon={filterIcon(key)}
-                      value={appliedSummary(values)}
-                      removable={false}
-                    />
-                  );
-                }
-                case "statuses": {
-                  const values = Array.isArray(value) ? value : [`${value}`];
-                  return (
-                    <AppliedFilter
-                      variant="minimal/medium"
-                      key={key}
-                      label={filterTitle(key)}
-                      icon={filterIcon(key)}
-                      value={appliedSummary(values.map((v) => runStatusTitle(v as TaskRunStatus)))}
-                      removable={false}
-                    />
-                  );
-                }
-                case "tags": {
-                  const values = Array.isArray(value) ? value : [`${value}`];
-                  return (
-                    <AppliedFilter
-                      variant="minimal/medium"
-                      key={key}
-                      label={filterTitle(key)}
-                      icon={filterIcon(key)}
-                      value={appliedSummary(values)}
-                      removable={false}
-                    />
-                  );
-                }
-                case "bulkId": {
-                  return (
-                    <AppliedFilter
-                      variant="minimal/medium"
-                      key={key}
-                      label={filterTitle(key)}
-                      icon={filterIcon(key)}
-                      value={value}
-                      removable={false}
-                    />
-                  );
-                }
-                case "rootOnly": {
-                  return (
-                    <AppliedFilter
-                      variant="minimal/medium"
-                      key={key}
-                      label={"Root only"}
-                      icon={filterIcon(key)}
-                      value={
-                        value ? (
-                          <CheckIcon className="size-4" />
-                        ) : (
-                          <XCircleIcon className="size-4" />
-                        )
-                      }
-                      removable={false}
-                    />
-                  );
-                }
-                case "runId": {
-                  return (
-                    <AppliedFilter
-                      variant="minimal/medium"
-                      key={key}
-                      label={"Run ID"}
-                      icon={filterIcon(key)}
-                      value={value}
-                      removable={false}
-                    />
-                  );
-                }
-                case "batchId": {
-                  return (
-                    <AppliedFilter
-                      variant="minimal/medium"
-                      key={key}
-                      label={"Batch ID"}
-                      icon={filterIcon(key)}
-                      value={value}
-                      removable={false}
-                    />
-                  );
-                }
-                case "scheduleId": {
-                  return (
-                    <AppliedFilter
-                      variant="minimal/medium"
-                      key={key}
-                      label={"Schedule ID"}
-                      icon={filterIcon(key)}
-                      value={value}
-                      removable={false}
-                    />
-                  );
-                }
-                default: {
-                  assertNever(typedKey);
-                }
-              }
-            })}
-          </div>
-        </div>
-      );
-    }
-  }
-}
-
-function Action({ action }: { action: BulkActionAction }) {
-  switch (action) {
-    case "cancel":
-      return (
-        <span>
-          <XCircleIcon className="mb-0.5 inline-block size-4 text-error" />
-          <span className="ml-0.5 text-text-bright">Canceled</span>
-        </span>
-      );
-    case "replay":
-      return (
-        <span>
-          <ArrowPathIcon className="mb-0.5 inline-block size-4 text-blue-400" />
-          <span className="ml-0.5 text-text-bright">Replayed</span>
-        </span>
-      );
-  }
-}
-
-function EstimatedCount({ count }: { count?: number }) {
-  if (typeof count === "number") {
-    return <>~{formatNumber(count)}</>;
-  }
-
-  return <SpinnerWhite className="mx-0.5 -mt-0.5 inline size-3" />;
 }
